requirements.txt
Pillow==11.1.0
click==8.1.8
numpy==2.2.5